requirements.txt
python-telegram-bot==21.3
google-genai==0.3.0
flask==3.0.3